fix: AuctionEntity auctionId 타입 Long-> Int 변경

+++ src/main/java/com/bidnbuy/server/controller/ChatRoomController.java
@@ -28,7 +28,7 @@ public class ChatRoomController {
     private final ChatMessageService chatMessageService;
 
     @PostMapping("/{auctionId}")
-    public ResponseEntity<ChatRoomDto> createChatRoom(@PathVariable Long auctionId,
+    public ResponseEntity<ChatRoomDto> createChatRoom(@PathVariable Integer auctionId,
                                                       @AuthenticationPrincipal Long userId,
                                                       @RequestBody ChatRoomCreateRequestDto requestDto){
         requestDto.setAuctionId(auctionId);
