+++ src/main/java/com/bidnbuy/server/dto/AuctionSalesHistoryDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Getter
 @Builder
 public class AuctionSalesHistoryDto {
-    private Long auctionId;
+    private Integer auctionId;
     private String title;
     private String itemImageUrl;
